// shared/types.ts
export interface RegistrationResponse {
  success: boolean;
  message: string;
  attendee: {
    id: string;
    name: string;
    email: string;
    ticketType: string;
    ticketCode: string;
    qrToken: string;
    qrDataUrl: string;
    registeredAt: string;
    emailSent: boolean;
  };
  event: {
    id: string;
    name: string;
    venue: string;
    date: string;
    startTime: string;
  };
}